Split API Client tests into separate modules for each client
The old module was way too long.
